Refactor backend endpoints
