+++ src/services/propertyService.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 
-const base_url = 'http://localhost:3001';
-const single_property_url = (id) => `http://localhost:3001/getSingleProperty/${id}`;
+const base_url = 'http://localhost:3001/api';
+const single_property_url = (id) => `${base_url}/getSingleProperty/${id}`;
 
 export const getAll = async (username) => {
     try {
@@ -35,7 +35,7 @@ export const singleProperty = async (id) => {
 
 export const createProperty = async (propertyData) => {
     try {
-      const response = await axios.post(`http://localhost:3001/createProperty`, propertyData);
+      const response = await axios.post(`${base_url}/createProperty`, propertyData);
       return response.data;
     } catch (error) {
       console.error("Error while creating property", error);
@@ -44,7 +44,7 @@ export const createProperty = async (propertyData) => {
   };
 
 export const updateProperty = async (id, data) => {
-    return axios.put(`http://localhost:3001/updateProperty/${id}`, data)
+    return axios.put(`${base_url}/updateProperty/${id}`, data)
         .then(response => response.data)
         .catch(error => {
             console.error('Error updating query: ', error );
@@ -53,7 +53,7 @@ export const updateProperty = async (id, data) => {
 }
 
 export const updateResident = (propertyId, { residentNumber }) => {
-    return axios.put(`http://localhost:3001/updateResident/${propertyId}`, {
+    return axios.put(`${base_url}/updateResident/${propertyId}`, {
       residentNumber
     })
     .then(response => response.data) // Връща данните от отговора, които могат да се използват във фронтенда
@@ -65,7 +65,7 @@ export const updateResident = (propertyId, { residentNumber }) => {
 
 export const deleteProperty = async (id) => {
     try {
-        const response = await axios.delete(`http://localhost:3001/deleteProperty/${id}`);
+        const response = await axios.delete(`${base_url}/deleteProperty/${id}`);
         return response.data;  // Return the response data
     } catch (error) {
         console.error('Error deleting property:', error);
@@ -75,7 +75,7 @@ export const deleteProperty = async (id) => {
 
 export const getAddressesPerUser = async (username) => {
     try {
-        const response = await fetch(`http://localhost:3001/getAllPropertiesPerUser?created_by=${username}`);
+        const response = await fetch(`${base_url}/getAllPropertiesPerUser?created_by=${username}`);
         if (!response.ok) {
             throw new Error('Failed to fetch addresses');
         }
